Refactor aws auth code out of DockerImagePackageDownloader
