+++ source/Calamari.CloudAccounts/Calamari.CloudAccounts.csproj
@@ -17,6 +17,7 @@
     <ItemGroup>
       <PackageReference Include="AWSSDK.Core" Version="3.7.400.62" />
       <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.401.11" />
+      <PackageReference Include="AWSSDK.ECR" Version="3.7.100.0" />
       <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
       <PackageReference Include="Microsoft.Identity.Client" Version="4.66.2" />
       <PackageReference Include="Microsoft.Rest.ClientRuntime" Version="2.3.24" />
