+++ source/Calamari.Shared/Calamari.Shared.csproj
@@ -37,6 +37,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Calamari.CloudAccounts\Calamari.CloudAccounts.csproj" />
     <ProjectReference Include="..\Calamari.Common\Calamari.Common.csproj" />
     <PackageReference Include="AWSSDK.S3" Version="3.7.8.8" />
     <PackageReference Include="BouncyCastle.Cryptography" Version="2.4.0" />
@@ -57,9 +58,6 @@
     <PackageReference Include="Microsoft.Web.Xdt" Version="3.1.0" />
     <PackageReference Include="System.Diagnostics.Tracing" Version="4.3.0" />
     <PackageReference Include="Polly" Version="8.3.1" />
-    <PackageReference Include="AWSSDK.Core" Version="3.7.400.62" />
-    <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.401.11" />
-    <PackageReference Include="AWSSDK.ECR" Version="3.7.100.0" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
     <PackageReference Include="NuGet.Commands" Version="5.11.5" />
